move BasePipeline to a new file
