+++ src/deepsparse/__init__.py
@@ -33,6 +33,7 @@ from .cpu import (
 from .engine import *
 from .tasks import *
 from .pipeline import *
+from .base_pipeline import *
 from .loggers import *
 from .version import __version__, is_release
 from .analytics import deepsparse_analytics as _analytics
